requirements.txt
Flask
flask-socketio
torch
numpy
pandas
requests
Pillow
eventlet
matplotlib